Moved external sources to src/ext/
