+++ CMakeLists.txt
@@ -254,7 +254,7 @@ set(CORE_SOURCES
   src/FontCache.cc
   src/DrawingCallback.cc
   src/FreetypeRenderer.cc
-  src/lodepng.cpp
+  src/ext/lodepng/lodepng.cpp
   src/PlatformUtils.cc 
   src/libsvg/circle.cc
   src/libsvg/ellipse.cc
@@ -272,7 +272,7 @@ set(CORE_SOURCES
   src/clipper-utils.cc
   src/assignment.cc
   src/annotation.cc 
-  src/polyclipping/clipper.cpp
+  src/ext/polyclipping/clipper.cpp
   ${PLATFORM_SOURCES}
   ${FLEX_openscad_lexer_OUTPUTS}
   ${BISON_openscad_parser_OUTPUTS}
@@ -314,20 +314,20 @@ set(CGAL_SOURCES
   src/svg.cc
   src/GeometryEvaluator.cc)
 
-include_directories("src/libtess2/Include")
+include_directories("src/ext/libtess2/Include")
 set(COMMON_SOURCES
   src/nodedumper.cc 
   src/GeometryCache.cc 
   src/clipper-utils.cc 
   src/Tree.cc
-  src/polyclipping/clipper.cpp
-  src/libtess2/Source/bucketalloc.c
-  src/libtess2/Source/dict.c
-  src/libtess2/Source/geom.c
-  src/libtess2/Source/mesh.c
-  src/libtess2/Source/priorityq.c
-  src/libtess2/Source/sweep.c
-  src/libtess2/Source/tess.c
+  src/ext/polyclipping/clipper.cpp
+  src/ext/libtess2/Source/bucketalloc.c
+  src/ext/libtess2/Source/dict.c
+  src/ext/libtess2/Source/geom.c
+  src/ext/libtess2/Source/mesh.c
+  src/ext/libtess2/Source/priorityq.c
+  src/ext/libtess2/Source/sweep.c
+  src/ext/libtess2/Source/tess.c
   src/Tree.cc)
 
 #
