+++ src/GeometryUtils.cc
@@ -1,5 +1,5 @@
 #include "GeometryUtils.h"
-#include "tesselator.h"
+#include "ext/libtess2/Include/tesselator.h"
 #include "printutils.h"
 #include "Reindexer.h"
 #include <boost/lexical_cast.hpp>
