+++ src/ext/libtess2/Source/dict.c
