+++ src/ext/libtess2/Source/mesh.h
