+++ src/ext/libtess2/Source/sweep.c
