+++ src/ext/libtess2/Source/tess.h
