+++ src/ext/lodepng/lodepng.cpp
