+++ src/ext/polyclipping/clipper.cpp
