+++ src/ext/polyclipping/clipper.hpp
